New metrics and result files

+++ runtime-EclipseApplication/TTSTestProject/src/metrics/custom/cellTimeMetric.java
@@ -75,7 +75,7 @@ public class cellTimeMetric extends Metric {
 		String topic = msg.getTopic();
 		if (topic.contains(completionTopicName) && topicMatches(topic)) {
 			SafetyZone sv = (SafetyZone)msg.getValue();
-			float level = sv.getLevel();
+			double level = sv.getLevel();
 			String object1 = sv.getObject1();
 			String object2 = sv.getObject2();
 			String zoneID = sv.getZoneID();
