+++ runtime-EclipseApplication/TTSTestProject/src/metrics/custom/combinedCollisionMetric.java
@@ -38,7 +38,7 @@ public abstract class combinedCollisionMetric extends BatchedRateMetric {
 		String topic = msg.getTopic();
 		if (topic.contains(completionTopicName) && topicMatches(topic)) {
 			SafetyZone sv = (SafetyZone) msg.getValue();
-			float level = sv.getLevel();
+			double level = sv.getLevel();
 			String object1 = sv.getObject1();
 			String object2 = sv.getObject2();
 			String zoneID = sv.getZoneID();
